+++ CHANGELOG.md
@@ -3,3 +3,6 @@ All Notable changes to bitsika binance-sdk will be documented in this file
 
 ## 2021-09-27
 - Initial release
+
+## 2022-03-23
+- Transaction receipt